+++ packages/reactabular/CHANGELOG.md
@@ -5,6 +5,7 @@
   * Improvement - reactabular - New `search-field` module for searching (UI).
   * Improvement - table - Make `rowKey` propType check compatible with React 15.3. It should give you better output during development now.
   * Improvement - easy - Expose `headerExtra` prop. It can be used to inject extra rows to a header. This works well with `reactabular-search-columns`.
+  * Improvement - easy - Expose `onSort` and `sortingColumns` props. These are useful for implementing sorting persistency.
 
 1.2.2 / 2016-08-07
 ==================